Make sure test packages attaches name to arg spies
